### 2019-03-12 #1316 @whedon
[REVIEW]: Badread: simulation of error-prone long reads
**Submitting author:** @rrwick (<a href="http://orcid.org/0000-0001-8349-0778">Ryan Wick</a>)
**Repository:** <a href="https://github.com/rrwick/Badread" target ="_blank">https://github.com/rrwick/Badread</a>
**Version:** v0.1.3
