### 2019-03-20 rrwick/Badread#1 @betteridiot
Address Community guidelines
With respect to https://github.com/openjournals/joss-reviews/issues/1316:
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Therefore, please take a moment to make a [CONTRIBUTING.md](https://help.github.com/en/articles/setting-guidelines-for-repository-contributors) file in your repo.

Secondly, it might be prudent to set up issue/bug templates in GitHub. This can be done easily by navigating to your repo's Settings > Options > Issues > Set up templates button.

Also, consider adding a CODE_OF_CONDUCT.md file (GitHub will provide some templates if you follow this [guide](https://help.github.com/en/articles/adding-a-code-of-conduct-to-your-project))
