### 2019-03-25 c38327e
Repair bug-fix issue template
