+++ .github/ISSUE_TEMPLATE/bug-report.md
@@ -1,24 +1,0 @@
----
-name: Bug report
-about: Create a report to help us improve
-title: ''
-labels: bug
-assignees: ''
-
----
-
-**Describe the bug**
-A clear and concise description of what the bug is.
-
-**To reproduce**
-The full command run which produces the error. If possible, attach an input files so I can try reproduce the problem on my end.
-
-**Error message**
-If applicable, include the full Python error message which indicates where the crash occurred.
-
-**System info**
- - OS: [e.g. Ubuntu 16.04]
- - Python version [e.g. 3.6.7]
-
-**Additional context**
-Add any other context about the problem here.
