### 2019-03-25 rrwick/Badread#1 @rrwick
Thanks for the suggestions! I've set up appropriate issue templates and added a [CONTRIBUTING.md](https://github.com/rrwick/Badread/blob/master/CONTRIBUTING.md) file. Instead of making a separate CODE_OF_CONDUCT.md file, I opted to include the code of conduct guidelines within the CONTRIBUTING.md file - let me know if you think that's okay.

Ryan
