### 2019-03-25 rrwick/Badread#2 @betteridiot
JoSS Review Notes
With respect to https://github.com/openjournals/joss-reviews/issues/1316: here are my notes for my complete review.

The only thing to notes is the issue with some machines not having the most up-to-date version of the `build-essentials` because of `edlib`'s requirement for `c++` and changing how `pip` should install. For 1) Having a FAQ page might be able to preempt this problem in the future and 2) just don't suggest users invoke `pip3`, just use `pip`.

This is **used only for review purposes**: the minor issues raised will have individual issues generated.

---
# JoSS Review pipeline for [Badread](https://github.com/rrwick/Badread)

---
# Documentation:

* CLI documentation present in README.md
* Detailed explanations of the various parameters present in the GH Wiki
* Use of Desmos plot page to explain the Gamma in terms of read/fragment length was terrific

## Documentation Evaluation:
Excellent documentation! While the norm tends to be a Readthedocs site, the use
of GitHub's Wiki was a great way to consolidate the details of the program.

---
# Installation:
Badread states that it requires Python >= 3.6

## Current `conda` version:
```bash
(base) $ conda --version
conda 4.6.8
```

## Setup environment:
```bash
(base) $ conda create -n badread python=3.6 --no-default-packages && conda activate badread
(badread) $ python --version
Python 3.6.7
```

## Install `badread` based on instructions:
```bash
(badread) $ pip3 install git+https://github.com/rrwick/Badread.git
Command 'pip3' not found ...
```

## Install using built-in environment `pip`:
```bash
(badread) $ pip install git+https://github.com/rrwick/Badread.git
(badread) ~ # pip install git+https://github.com/rrwick/Badread.git
Collecting git+https://github.com/rrwick/Badread.git
  Cloning https://github.com/rrwick/Badread.git to /tmp/pip-req-build-6sj3au9h
Collecting edlib (from Badread==0.1.3)
  Downloading https://files.pythonhosted.org/packages/33/04/0dd73c02c0a5148754a6144df9a7b5dd55ccdbb1b17481bbc89b0193c544/edlib-1.2.4.tar.gz (54kB)
    100% |████████████████████████████████| 61kB 1.3MB/s
Collecting numpy (from Badread==0.1.3)
  Downloading https://files.pythonhosted.org/packages/35/d5/4f8410ac303e690144f0a0603c4b8fd3b986feb2749c435f7cdbb288f17e/numpy-1.16.2-cp36-cp36m-manylinux1_x86_64.whl (17.3MB)
    100% |████████████████████████████████| 17.3MB 1.4MB/s
Collecting scipy (from Badread==0.1.3)
  Downloading https://files.pythonhosted.org/packages/7f/5f/c48860704092933bf1c4c1574a8de1ffd16bf4fde8bab190d747598844b2/scipy-1.2.1-cp36-cp36m-manylinux1_x86_64.whl (24.8MB)
    100% |████████████████████████████████| 24.8MB 795kB/s
Building wheels for collected packages: Badread, edlib
  Building wheel for Badread (setup.py) ... done
  Stored in directory: /tmp/pip-ephem-wheel-cache-p2cgzrv1/wheels/db/38/bd/d192f5ff0b1d97bab8a1b7478730232764f120170c8a730e57
  Building wheel for edlib (setup.py) ... error
  Complete output from command /root/miniconda3/envs/badread/bin/python -u -c "import setuptools, tokenize;__file__='/tmp/pip-install-926py29e/edlib/setup.py';f=getattr(tokenize, 'open', open)(__file__);code=f.read().replace('\r\n', '\n');f.close();exec(compile(code, __file__, 'exec'))" bdist_wheel -d /tmp/pip-wheel-jd40pj7t --python-tag cp36:
  running bdist_wheel
  running build
  running build_ext
  building 'edlib' extension
  creating build
  creating build/temp.linux-x86_64-3.6
  creating build/temp.linux-x86_64-3.6/edlib
  creating build/temp.linux-x86_64-3.6/edlib/src
  gcc -pthread -B /root/miniconda3/envs/badread/compiler_compat -Wl,--sysroot=/ -Wsign-compare -DNDEBUG -g -fwrapv -O3 -Wall -Wstrict-prototypes -fPIC -Iedlib/include -I/root/miniconda3/envs/badread/include/python3.6m -c edlib.bycython.cpp -o build/temp.linux-x86_64-3.6/edlib.bycython.o -O3 -std=c++11
  gcc: error trying to exec 'cc1plus': execvp: No such file or directory
  error: command 'gcc' failed with exit status 1

  Failed building wheel for edlib
  Running setup.py clean for edlib
Successfully built Badread
Failed to build edlib
Installing collected packages: edlib, numpy, scipy, Badread
  Running setup.py install for edlib ... error
    Complete output from command /root/miniconda3/envs/badread/bin/python -u -c "import setuptools, tokenize;__file__='/tmp/pip-install-926py29e/edlib/setup.py';f=getattr(tokenize, 'open', open)(__file__);code=f.read().replace('\r\n', '\n');f.close();exec(compile(code, __file__, 'exec'))" install --record /tmp/pip-record-b2nec6ur/install-record.txt --single-version-externally-managed --compile:
    running install
    running build
    running build_ext
    building 'edlib' extension
    creating build
    creating build/temp.linux-x86_64-3.6
    creating build/temp.linux-x86_64-3.6/edlib
    creating build/temp.linux-x86_64-3.6/edlib/src
    gcc -pthread -B /root/miniconda3/envs/badread/compiler_compat -Wl,--sysroot=/ -Wsign-compare -DNDEBUG -g -fwrapv -O3 -Wall -Wstrict-prototypes -fPIC -Iedlib/include -I/root/miniconda3/envs/badread/include/python3.6m -c edlib.bycython.cpp -o build/temp.linux-x86_64-3.6/edlib.bycython.o -O3 -std=c++11
    gcc: error trying to exec 'cc1plus': execvp: No such file or directory
    error: command 'gcc' failed with exit status 1

Command "/root/miniconda3/envs/badread/bin/python -u -c "import setuptools, tokenize;__file__='/tmp/pip-install-926py29e/edlib/setup.py';f=getattr(tokenize, 'open', open)(__file__);code=f.read().replace('\r\n', '\n');f.close();exec(compile(code, __file__, 'exec'))" install --record /tmp/pip-record-b2nec6ur/install-record.txt --single-version-externally-managed --compile" failed with error code 1 in /tmp/pip-install-926py29e/edlib/
```

## Installing requirements separately:
```bash
(badread) $ pip install numpy scipy matplotlib
(badread) $ pip install edlib
Collecting edlib
  Using cached https://files.pythonhosted.org/packages/33/04/0dd73c02c0a5148754a6144df9a7b5dd55ccdbb1b17481bbc89b0193c544/edlib-1.2.4.tar.gz
Building wheels for collected packages: edlib
  Building wheel for edlib (setup.py) ... error
  Complete output from command /root/miniconda3/envs/badread/bin/python -u -c "import setuptools, tokenize;__file__='/tmp/pip-install-9iupvpzl/edlib/setup.py';f=getattr(tokenize, 'open', open)(__file__);code=f.read().replace('\r\n', '\n');f.close();exec(compile(code, __file__, 'exec'))" bdist_wheel -d /tmp/pip-wheel-mdy8y0j1 --python-tag cp36:
  running bdist_wheel
  running build
  running build_ext
  building 'edlib' extension
  creating build
  creating build/temp.linux-x86_64-3.6
  creating build/temp.linux-x86_64-3.6/edlib
  creating build/temp.linux-x86_64-3.6/edlib/src
  gcc -pthread -B /root/miniconda3/envs/badread/compiler_compat -Wl,--sysroot=/ -Wsign-compare -DNDEBUG -g -fwrapv -O3 -Wall -Wstrict-prototypes -fPIC -Iedlib/include -I/root/miniconda3/envs/badread/include/python3.6m -c edlib.bycython.cpp -o build/temp.linux-x86_64-3.6/edlib.bycython.o -O3 -std=c++11
  gcc: error trying to exec 'cc1plus': execvp: No such file or directory
  error: command 'gcc' failed with exit status 1

  Failed building wheel for edlib
  Running setup.py clean for edlib
Failed to build edlib
Installing collected packages: edlib
  Running setup.py install for edlib ... error
    Complete output from command /root/miniconda3/envs/badread/bin/python -u -c "import setuptools, tokenize;__file__='/tmp/pip-install-9iupvpzl/edlib/setup.py';f=getattr(tokenize, 'open', open)(__file__);code=f.read().replace('\r\n', '\n');f.close();exec(compile(code, __file__, 'exec'))" install --record /tmp/pip-record-0oaizsc1/install-record.txt --single-version-externally-managed --compile:
    running install
    running build
    running build_ext
    building 'edlib' extension
    creating build
    creating build/temp.linux-x86_64-3.6
    creating build/temp.linux-x86_64-3.6/edlib
    creating build/temp.linux-x86_64-3.6/edlib/src
    gcc -pthread -B /root/miniconda3/envs/badread/compiler_compat -Wl,--sysroot=/ -Wsign-compare -DNDEBUG -g -fwrapv -O3 -Wall -Wstrict-prototypes -fPIC -Iedlib/include -I/root/miniconda3/envs/badread/include/python3.6m -c edlib.bycython.cpp -o build/temp.linux-x86_64-3.6/edlib.bycython.o -O3 -std=c++11
    gcc: error trying to exec 'cc1plus': execvp: No such file or directory
    error: command 'gcc' failed with exit status 1

Command "/root/miniconda3/envs/badread/bin/python -u -c "import setuptools, tokenize;__file__='/tmp/pip-install-9iupvpzl/edlib/setup.py';f=getattr(tokenize, 'open', open)(__file__);code=f.read().replace('\r\n', '\n');f.close();exec(compile(code, __file__, 'exec'))" install --record /tmp/pip-record-0oaizsc1/install-record.txt --single-version-externally-managed --compile" failed with error code 1 in /tmp/pip-install-9iupvpzl/edlib/
(badread) $ apt-get install --reinstall build-essential
(badread) $ pip install edlib
(badread) $ pip install git+https://github.com/rrwick/Badread.git
(badread) $ badread --help
usage: badread [-h] [--version] {simulate,error_model,qscore_model,plot} ...

Badread: a long read simulator that can imitate manytypes of read problems

Commands:
  {simulate,error_model,qscore_model,plot}
                              simulate:     generate fake long reads
                              error_model:  build a Badread error model
                              qscore_model: build a Badread qscore model
                              plot:         view read identities over a sliding window

Help:
  -h, --help                  Show this help message and exit
  --version                   Show program's version number and exit
```

## Installation Evaluation:
Because of `edlib`, you should state that an additional requirement of install
is to have an updated version of `build-essential`.

---
# Quickstart:
Using the [*Bacillus subtilis*](ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/000/009/045/GCF_000009045.1_ASM904v1/GCF_000009045.1_ASM904v1_genomic.fna.gz)
from [NCBI](https://www.ncbi.nlm.nih.gov/genome/?term=Bacillus%20subtilis[Organism]&cmd=DetailsSearch)

## Create "normal" reads:
```bash
(badread) $ badread simulate --reference data/ref.fna --quantity 25x | gzip > data/reads.fastq.gz

Badread v0.1.3
long read simulation

Loading reference from ref.fna
  1 contig:
    NC_000964.3: 4,215,606 bp, linear, 1.0x depth

Generating fragment lengths from a gamma distribution:
  mean  =  15000 bp      parameters:
  stdev =  13000 bp        k (shape)     = 1.3314e+00
  N50   =  22622 bp        theta (scale) = 1.1267e+04
  │ ▖▌▌▖
  │ ▌▌▌▌▌▌▖
f │▌▌▌▌▌▌▌▌▌▖
r │▌▌▌▌▌▌▌▌▌▌▌▖
a │▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖
g │▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖
s │▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖
  │▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖▖▖▖▖▖▖
  ├─────────┬─────────┬─────────┬─────────┬─────────┐
  │       ▖▖▌▌▌▌▌▖▖
  │      ▌▌▌▌▌▌▌▌▌▌▌▌▖
b │    ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖
a │   ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖
s │   ▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖
e │  ▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖
s │ ▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖
  │▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖
  ├─────────┬─────────┬─────────┬─────────┬─────────┐
  0         13600     27200     40800     54400     68000

Generating read identities from a beta distribution:
  mean  =  85%      shape parameters:
  max   =  95%        alpha = 2.9526e+01
  stdev =   5%        beta  = 3.4737e+00
  │                                   ▖▌▌▌
  │                                  ▌▌▌▌▌▌
  │                                ▖▌▌▌▌▌▌▌▌
  │                               ▖▌▌▌▌▌▌▌▌▌▖
  │                              ▌▌▌▌▌▌▌▌▌▌▌▌
  │                            ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▖
  │                         ▖▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌
  │                    ▖▖▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌
  ├─────────┬─────────┬─────────┬─────────┬─────────┐
  50        60        70        80        90        100

Loading error model from <MINICONDAPATH>/site-packages/badread/error_models/nanopore.gz
  done: loaded error distributions for 16384 7-mers

Loading qscore model from <MINICONDAPATH>/site-packages/badread/qscore_models/nanopore.gz
  done: loaded qscore distributions for 19458 alignments

Read glitches:
  rate (mean distance between glitches) = 10000
  size (mean length of random sequence) =    25
  skip (mean sequence lost per glitch)  =    25

Start adapter:
  seq: AATGTACTTCGTTCAGTTACGTATTGCT
  rate:   90.0%
  amount: 60.0%

End adapter:
  seq: GCAATACGTAACTGAACGAAGT
  rate:   50.0%
  amount: 20.0%

Other problems:
  chimera join rate: 1%
  junk read rate:    1%
  random read rate:  1%

Target read set size: 105,390,150 bp

Simulating: 7,169 reads  105,427,004 bp  100.0%
```

## Create pacbio reads:

```bash
(badread) $ badread simulate --reference data/ref.fna --quantity 25x --error_model pacbio
--qscore_model pacbio --identity 85,95,3 --length 7500,7500 |gzip > data/pb_reads.fastq.gz

Badread v0.1.3
long read simulation

Loading reference from ref.fna
  1 contig:
    NC_000964.3: 4,215,606 bp, linear, 1.0x depth

Generating fragment lengths from a gamma distribution:
  mean  =   7500 bp      parameters:
  stdev =   7500 bp        k (shape)     = 1.0000e+00
  N50   =  12588 bp        theta (scale) = 7.5000e+03
  │▌▖
  │▌▌▌▌▖
f │▌▌▌▌▌▌▌▖▖
r │▌▌▌▌▌▌▌▌▌▌▌▖▖
a │▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖
g │▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖
s │▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖▖▖▖
  │▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖▖▖▖▖▖▖▖▖▖▖▖▖▖▖▖▖▖▖
  ├─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
  │            ▖▖▖▌▌▌▌▌▌▌▌▌▌▖▖▖▖▖
  │         ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖
b │       ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖
a │     ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖
s │    ▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖
e │  ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖▖▖
s │ ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖▖▖▖▖▖
  │▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖
  ├─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
  0         3800      7600      11400     15200     19000     22800     26600     30400     34200     38000

Generating read identities from a beta distribution:
  mean  =  85%      shape parameters:
  max   =  95%        alpha = 8.3608e+01
  stdev =   3%        beta  = 9.8363e+00
  │                                                                     ▖▌▌▌▖
  │                                                                   ▖▌▌▌▌▌▌▌
  │                                                                  ▌▌▌▌▌▌▌▌▌▌▖
  │                                                                ▖▌▌▌▌▌▌▌▌▌▌▌▌▖
  │                                                               ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌
  │                                                             ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖
  │                                                          ▖▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖
  │                                                    ▖▖▖▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖
  ├─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
  50        55        60        65        70        75        80        85        90        95        100

Loading error model from <MINICONDAPATH>/site-packages/badread/error_models/pacbio.gz
  done: loaded error distributions for 16384 7-mers

Loading qscore model from <MINICONDAPATH>/site-packages/badread/qscore_models/pacbio.gz
  done: loaded qscore distributions for 8291 alignments

Read glitches:
  rate (mean distance between glitches) = 10000
  size (mean length of random sequence) =    25
  skip (mean sequence lost per glitch)  =    25

Start adapter:
  seq: AATGTACTTCGTTCAGTTACGTATTGCT
  rate:   90.0%
  amount: 60.0%

End adapter:
  seq: GCAATACGTAACTGAACGAAGT
  rate:   50.0%
  amount: 20.0%

Other problems:
  chimera join rate: 1%
  junk read rate:    1%
  random read rate:  1%

Target read set size: 105,390,150 bp

Simulating: 14,116 reads  105,406,061 bp  100.0%
```

## Create "bad" reads:

```bash
(badread) $ badread simulate --reference data/ref.fna --quantity 25x --glitches 1000,100
,100 --junk_reads 5 --random_reads 5 --chimeras 10 --identity 75,90,8 | gzip > data/bad_reads.fastq.gz

Badread v0.1.3
long read simulation

Loading reference from ref.fna
  1 contig:
    NC_000964.3: 4,215,606 bp, linear, 1.0x depth

Generating fragment lengths from a gamma distribution:
  mean  =  15000 bp      parameters:
  stdev =  13000 bp        k (shape)     = 1.3314e+00
  N50   =  22622 bp        theta (scale) = 1.1267e+04
  │ ▖▌▌▖
  │ ▌▌▌▌▌▌▖
f │▌▌▌▌▌▌▌▌▌▖
r │▌▌▌▌▌▌▌▌▌▌▌▖
a │▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖
g │▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖
s │▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖
  │▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖▖▖▖▖▖▖
  ├─────────┬─────────┬─────────┬─────────┬─────────┐
  │       ▖▖▌▌▌▌▌▖▖
  │      ▌▌▌▌▌▌▌▌▌▌▌▌▖
b │    ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖
a │   ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖
s │   ▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖
e │  ▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖
s │ ▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖
  │▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖
  ├─────────┬─────────┬─────────┬─────────┬─────────┐
  0         13600     27200     40800     54400     68000

Generating read identities from a beta distribution:
  mean  =  75%      shape parameters:
  max   =  90%        alpha = 1.3815e+01
  stdev =   8%        beta  = 2.7630e+00
  │                         ▖▖▌▌▌▌▖
  │                       ▖▌▌▌▌▌▌▌▌▌▖
  │                     ▖▌▌▌▌▌▌▌▌▌▌▌▌▖
  │                   ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖
  │                ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖
  │             ▖▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌
  │         ▖▖▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌
  │ ▖▖▖▖▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌
  ├─────────┬─────────┬─────────┬─────────┬─────────┐
  50        60        70        80        90        100

Loading error model from <MINICONDAPATH>/site-packages/badread/error_models/nanopore.gz
  done: loaded error distributions for 16384 7-mers

Loading qscore model from <MINICONDAPATH>/site-packages/badread/qscore_models/nanopore.gz
  done: loaded qscore distributions for 19458 alignments

Read glitches:
  rate (mean distance between glitches) =  1000
  size (mean length of random sequence) =   100
  skip (mean sequence lost per glitch)  =   100

Start adapter:
  seq: AATGTACTTCGTTCAGTTACGTATTGCT
  rate:   90.0%
  amount: 60.0%

End adapter:
  seq: GCAATACGTAACTGAACGAAGT
  rate:   50.0%
  amount: 20.0%

Other problems:
  chimera join rate: 10.0%
  junk read rate:    5.0%
  random read rate:  5.0%

Target read set size: 105,390,150 bp

Simulating: 6,967 reads  105,392,600 bp  100.0%
```

## Create "good" reads:
```bash
(badread) $ badread simulate --reference data/ref.fna --quantity 25x --error_model random
--qscore_model ideal --glitches 0,0,0 --junk_reads 0 --random_reads 0 --chimeras 0
--identity 95,100,4 --start_adapter_seq "" --end_adapter_seq "" | gzip > data/good_reads.fastq.gz

Badread v0.1.3
long read simulation

Loading reference from ref.fna
  1 contig:
    NC_000964.3: 4,215,606 bp, linear, 1.0x depth

Generating fragment lengths from a gamma distribution:
  mean  =  15000 bp      parameters:
  stdev =  13000 bp        k (shape)     = 1.3314e+00
  N50   =  22622 bp        theta (scale) = 1.1267e+04
  │ ▖▌▌▖
  │ ▌▌▌▌▌▌▖
f │▌▌▌▌▌▌▌▌▌▖
r │▌▌▌▌▌▌▌▌▌▌▌▖
a │▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖
g │▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖
s │▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖
  │▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖▖▖▖▖▖▖
  ├─────────┬─────────┬─────────┬─────────┬─────────┐
  │       ▖▖▌▌▌▌▌▖▖
  │      ▌▌▌▌▌▌▌▌▌▌▌▌▖
b │    ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖
a │   ▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖
s │   ▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖
e │  ▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖
s │ ▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖▖
  │▖▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▌▖▖▖
  ├─────────┬─────────┬─────────┬─────────┬─────────┐
  0         13600     27200     40800     54400     68000

Generating read identities from a beta distribution:
  mean  =  95%      shape parameters:
  max   = 100%        alpha = 2.7253e+01
  stdev =   4%        beta  = 1.4344e+00
  │                                               ▖▌
  │                                              ▖▌▌▖
[…]
