### 2019-03-25 rrwick/Badread#4 @betteridiot
Change install documentation
**Describe the bug**
`pip3` is not guaranteed to be on every system. Therefore the note to use `pip3` for install may cause install errors. See https://github.com/rrwick/Badread/issues/2 for all output.

**To reproduce**
```bash
(base) $ conda --version
conda 4.6.8
(base) $ conda create -n badread python=3.6 --no-default-packages && conda activate badread
(badread) $ python --version
Python 3.6.7
(badread) $ pip3 install git+https://github.com/rrwick/Badread.git
Command 'pip3' not found ...
```

**Error message**
See above

**System info**
 - OS: Ubuntu 18.04.2 LTS (Bionic Beaver) on Windows Subsystem for Linux
 - Python version Python 3.6.7
