### 2019-03-25 #1316 @betteridiot
Please check https://github.com/rrwick/Badread/issues/2 for my review process notes.

The minor issues raised were:
* https://github.com/rrwick/Badread/issues/3: Prompting users to update `build-essentials`
* https://github.com/rrwick/Badread/issues/1: Add community/contributions guidelines
* https://github.com/rrwick/Badread/issues/4: Change installation invocation within `Install` section of README.md

All in all, terrific work on the package: clean, organized, great test coverage, and excellent documentation.
