### 2019-03-25 #1316 @betteridiot
Once these *minor* issues are addressed, I would be more than happy to check the remaining boxes and give my support for publication.
