### 2019-04-01 d549d3e
Add reference to installation issues wiki page
