### 2019-04-01 rrwick/Badread#4 @rrwick
Thanks! Exact Python/pip commands can be tough as different ways of installing Python do things differently (e.g. pip vs pip2 vs pip3). I've added an [Installation issues](https://github.com/rrwick/Badread/wiki/Installation-issues) page to the wiki which covers all the possible pip problems I could think of (and compiler issues for #3). This page is linked to from the [installation section of the README](https://github.com/rrwick/Badread#installation).
