### 2019-04-01 rrwick/Badread#3 @rrwick
Thanks! It's easy to forget that not everyone has compilers set up on their machine :smile:

I've added an [Installation issues](https://github.com/rrwick/Badread/wiki/Installation-issues) page to the wiki and linked to it in the [installation section of the README](https://github.com/rrwick/Badread#installation). This page covers installing compilers for common OSs (and pip issues to address #4).
