### 2019-04-01 rrwick/Badread#4 @rrwick
I decided to leave the [installation commands](https://github.com/rrwick/Badread#install-from-source) in the README using `pip3` instead of just `pip`. This is mainly to guard against systems with both Python 2 and 3 installed, in which case `pip` probably means `pip2`.

But if you feel strongly that `pip` would be better, I can change them to that. I would then update the [installation issues wiki page](https://github.com/rrwick/Badread/wiki/Installation-issues) to cover what to do if users run into Python version errors. Let me know what you think.

I'll be happy when Python 2 breaths its last breath and we can all just assume Python == Python 3. It will help if and when macOS switches to Python 3 🤞
