### 2019-04-01 #1316 @rrwick
Thanks for the feedback! I think I've addressed all your points in the relevant issues pages on the Badread wiki.
