### 2019-04-01 rrwick/Badread#4 @betteridiot
No, you shouldn't need to change it. The fact that you added to the wiki is perfectly fine with me.

P.S. I am with you on the 2.7 thing.
