### 2019-04-01 rrwick/Badread#3 @betteridiot
Thank you. This addresses my concern here.
