### 2019-04-01 rrwick/Badread#2 @betteridiot
All concerns have been met. I will close this here and finish my review on https://github.com/openjournals/joss-reviews/issues/1316
