### 2019-04-03 #1316 @rrwick
Okay, I've made a Zenodo DOI (https://zenodo.org/record/2622997) and added this shiny new badge to Badread's README:
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.2622997.svg)](https://doi.org/10.5281/zenodo.2622997)
