### 2019-04-04 #1316 @kyleniemeyer
Congrats @rrwick on your JOSS article acceptance! Thanks to @betteridiot for reviewing and @brainstorm for editing.
